Create README.md
Our project delves into the intersection of emergency services and public health, focusing on the correlation between fire department calls and COVID testing data in the vibrant city of San Francisco. Leveraging the power of Hadoop, Oracle, and Linux servers, we conducted an in-depth analysis to glean insights that could contribute to optimizing emergency response programs in the region.